moved opensearch related stuff to external/elasticsearch pkg
